+++ chain/network/src/peer_manager/tests/mod.rs
@@ -1,0 +1,4 @@
+mod accounts_data;
+mod connection_pool;
+mod nonce;
+mod routing;
